Reorganize test classes
